+++ src/components/user/NoProducts.tsx
@@ -1,0 +1,9 @@
+export default function NoProducts() {
+	return (
+		<div className="p-[1rem] text-center md:p-[2rem] lg:p-[3rem]">
+			<span className="text-[1.7rem] font-semibold text-gray-200 lg:text-[1.9rem]">
+				상품이 없습니다.
+			</span>
+		</div>
+	);
+}
